requirements.txt
Flask
Gunicorn
python-gnupg
